feat[catalogPage]: implement all categories button (Sun, Sep 17, 2023 0:37)
