+++ src/assets/styles/components/_buttons.scss
@@ -69,6 +69,12 @@
     }
   }
 
+  &_active {
+    background-color: $color-light;
+    color: $color-primary;
+    border: 1px solid $color-primary;
+  }
+
   &:disabled {
     cursor: not-allowed;
     filter: grayscale(1);
