+++ src/assets/styles/components/_pagination.scss
@@ -8,7 +8,7 @@
   display: flex;
   gap: 10px;
 
-  > * {
+  >* {
     display: flex;
     align-items: center;
     justify-content: center;
@@ -16,12 +16,6 @@
     @media (max-width: 768px) {
       padding: 5px 10px;
     }
-  }
-
-  .active {
-    background-color: $color-light;
-    color: $color-primary;
-    border: 1px solid $color-primary;
   }
 
   @media (max-width: 530px) {
