+++ src/pages/catalogPage/catalogPageModel.ts
@@ -1,65 +1,6 @@
-/* eslint-disable no-console */
 import { getAnonymousFlow } from '../../shared/API';
-import { Product } from '../../types/interfaces/Product';
+import { Category, CategoryData, responseTypeCategory, responseTypeProduct } from '../../types/interfaces/catalogPage';
 import { generatePaginationForSelectedCategory, renderSelectedCategory } from './catalogPageController';
-
-type Category = {
-  ancestors?: string[];
-  assets?: string[];
-  createdAt: string;
-  createdBy: {
-    isPlatformClient: boolean;
-    user: {
-      id: string;
-      typeId: string;
-    };
-  };
-  description: Record<string, string>;
-  externalId: string;
-  id: string;
-  key: string;
-  lastMessageSequenceNumber: number;
-  lastModifiedAt: string;
-  lastModifiedBy: {
-    isPlatformClient: true;
-    user: {
-      id: string;
-      typeId: string;
-    };
-  };
-  metaDescription: Record<string, string>;
-  metaTitle: Record<string, string>;
-  name: Record<string, string>;
-  orderHint: string;
-  slug: Record<string, string>;
-  parent: {
-    typeId: string;
-    id: string;
-  };
-  version: number;
-  versionModifiedAt: string;
-};
-
-export type responseTypeProduct = {
-  count: number;
-  limit: number;
-  offset: number;
-  total: number;
-  results: Product[];
-};
-
-type responseTypeCategory = {
-  count: number;
-  limit: number;
-  offset: number;
-  total: number;
-  results: Category[];
-};
-
-type CategoryData = {
-  categoryName: string;
-  id: string;
-};
 
 export const LIMIT = 3;
 
@@ -67,7 +8,6 @@ export const getAllProducts = async (offset = 0): Promise<responseTypeProduct> =
   const path = `/products?limit=${LIMIT}&offset=${offset}`;
   const response: responseTypeProduct = await getAnonymousFlow(path);
   sessionStorage.setItem('productCount', `${response.total}`);
-  console.log(response.results[0]);
   return response;
 };
